# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Schema-register API v1",
    "version": "1.0.0",
    "contact": {
      "email": "developer.overheid@geonovum.nl",
      "name": "Team developer.overheid.nl",
      "url": "https://github.com/developer-overheid-nl/don-schema-register/issues"
    },
    "description": "API to access the Schema register of developer.overheid.nl.\n\n## Auth\n\nThis API distinguishes between public and private endpoints.\nPublic endpoints can be accessed with either an API key or a client credentials token.\nPrivate endpoints can only be accessed with a client credentials token.\n\n### API key\n\nUsing an API key, you can access all public endpoints of the Schema register.\nThese requests can also be made from the browser.\nRequest a read-only API key at https://apis.developer.overheid.nl/apis/key-aanvragen.\nSimply pass the obtained API key with each request using the `X-Api-Key` header.\n\n### Client credentials token\n\nUsing a client credentials token, you can access both public and private endpoints of the Schema register.\nTo obtain the token, you need to perform a `POST` request to `https://auth.developer.overheid.nl/realms/don/protocol/openid-connect/token` with the following Form URL Encoded body:\n- `grant_type`: `client_credentials`\n- `scope`: depending on the access you need and the client you are, you can request one or more of the following scopes:\n  - `schemas:read`\n  - `schemas:write`\n  - `organisations:read`\n  - `organisations:write`\n- `client_id`: the client id you received from us\n- `client_secret`: the client secret you received from us\n\nPass the obtained token with each request using the `Authorization` header. Example:\n\n`Authorization`: `Bearer {ACCESS_TOKEN}` (replace `{ACCESS_TOKEN}` with the obtained `access_token`)\n\n## Pagination\n\nPagination of collections is done using the `Link` header.\nThere are various libraries available (such as [parse-link-header](https://www.npmjs.com/package/parse-link-header) for Javascript) that can parse this header.\nAdditionally, the following headers provide extra support for implementing pagination in the client:\n- `Current-Page`: the current page in the collection\n- `Per-Page`: the number of items per page\n- `Total-Count`: the total number of items\n- `Total-Pages`: the total number of pages"
  },
  "servers": [
    {
      "description": "Production",
      "url": "https://api.developer.overheid.nl/schema-register/v1"
    },
    {
      "description": "Test",
      "url": "https://api.don.apps.digilab.network/schema-register/v1"
    }
  ],
  "paths": {
    "/schemas": {
      "get": {
        "tags": [
          "Public endpoints",
          "Schemas"
        ],
        "summary": "List schemas",
        "description": "Returns a list of JSON Schemas included in the register. Supports the same filter query parameters as the schema filter endpoint.",
        "operationId": "listSchemas",
        "parameters": [
          {
            "$ref": "#/components/parameters/Page"
          },
          {
            "$ref": "#/components/parameters/PerPage"
          },
          {
            "$ref": "#/components/parameters/SearchFilter"
          },
          {
            "$ref": "#/components/parameters/OrganisationFilter"
          },
          {
            "$ref": "#/components/parameters/DialectFilter"
          },
          {
            "$ref": "#/components/parameters/RootTypeFilter"
          }
        ],
        "responses": {
          "200": {
            "headers": {
              "API-Version": {
                "$ref": "#/components/headers/APIVersion"
              },
              "Link": {
                "$ref": "#/components/headers/Link"
              },
              "Total-Count": {
                "$ref": "#/components/headers/TotalCount"
              },
              "Current-Page": {
                "$ref": "#/components/headers/CurrentPage"
              },
              "Per-Page": {
                "$ref": "#/components/headers/PerPage"
              },
              "Total-Pages": {
                "$ref": "#/components/headers/TotalPages"
              }
            },
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SchemaSummary"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      },
      "post": {
        "tags": [
          "Private endpoints",
          "Schemas"
        ],
        "summary": "Create schema",
        "description": "Register a new JSON Schema in the register. The schema is fetched from schemaUrl (or read from schemaBody), validated and stored together with derived metadata (title, description, dialect, root type and a content hash).",
        "operationId": "createSchema",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SchemaPost"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "headers": {
              "API-Version": {
                "$ref": "#/components/headers/APIVersion"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SchemaSummary"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        },
        "security": [
          {
            "clientCredentials": [
              "schemas:write"
            ]
          }
        ]
      }
    },
    "/schemas/filters": {
      "get": {
        "tags": [
          "Public endpoints",
          "Schemas"
        ],
        "summary": "List schema filters",
        "description": "Returns all available filter options with counts. Counts are calculated based on the active filters and the optional search term q.",
        "operationId": "listSchemaFilters",
        "parameters": [
          {
            "$ref": "#/components/parameters/SearchFilter"
          },
          {
            "$ref": "#/components/parameters/OrganisationFilter"
          },
          {
            "$ref": "#/components/parameters/DialectFilter"
          },
          {
            "$ref": "#/components/parameters/RootTypeFilter"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "headers": {
              "API-Version": {
                "$ref": "#/components/headers/APIVersion"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FilterGroup"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
    "/schemas/{id}": {
      "get": {
        "tags": [
          "Public endpoints",
          "Schemas"
        ],
        "summary": "Get schema by id",
        "description": "Returns a single JSON Schema based on its id, including the stored schema content.",
        "operationId": "getSchemaById",
        "parameters": [
          {
            "$ref": "#/components/parameters/ResourceId"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "headers": {
              "API-Version": {
                "$ref": "#/components/headers/APIVersion"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SchemaDetail"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "put": {
        "tags": [
          "Private endpoints",
          "Schemas"
        ],
        "summary": "Update schema",
        "description": "Re-fetches the JSON Schema from schemaUrl (or reads it from schemaBody) and updates the registered schema. The organisationUri must match the owner of the schema.",
        "operationId": "updateSchema",
        "parameters": [
          {
            "$ref": "#/components/parameters/ResourceId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SchemaPost"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "headers": {
              "API-Version": {
                "$ref": "#/components/headers/APIVersion"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SchemaSummary"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "security": [
          {
            "clientCredentials": [
              "schemas:write"
            ]
          }
        ]
      }
    },
    "/schemas/{id}/schema.json": {
      "get": {
        "tags": [
          "Public endpoints",
          "Schemas"
        ],
        "summary": "Get raw schema content",
        "description": "Returns the raw stored JSON Schema document.",
        "operationId": "getSchemaContent",
        "parameters": [
          {
            "$ref": "#/components/parameters/ResourceId"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "headers": {
              "API-Version": {
                "$ref": "#/components/headers/APIVersion"
              }
            },
            "content": {
              "application/schema+json": {
                "schema": {
                  "type": "object",
                  "description": "The stored JSON Schema document"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/organisations": {
      "get": {
        "security": [
          {
            "apiKey": []
          },
          {
            "clientCredentials": [
              "organisations:read"
            ]
          }
        ],
        "tags": [
          "Public endpoints",
          "Organisations"
        ],
        "summary": "List organisations",
        "description": "List organisations",
        "operationId": "listOrganisations",
        "responses": {
          "200": {
            "headers": {
              "API-Version": {
                "$ref": "#/components/headers/APIVersion"
              },
              "Link": {
                "$ref": "#/components/headers/Link"
              },
              "Total-Count": {
                "$ref": "#/components/headers/TotalCount"
              },
              "Current-Page": {
                "$ref": "#/components/headers/CurrentPage"
              },
              "Per-Page": {
                "$ref": "#/components/headers/PerPage"
              },
              "Total-Pages": {
                "$ref": "#/components/headers/TotalPages"
              }
            },
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OrganisationSummary"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/Page"
          },
          {
            "$ref": "#/components/parameters/PerPage"
          }
        ]
      },
      "post": {
        "security": [
          {
            "clientCredentials": [
              "organisations:write"
            ]
          }
        ],
        "tags": [
          "Private endpoints",
          "Organisations"
        ],
        "summary": "Create organisation",
        "description": "Create a new organisation.",
        "operationId": "createOrganisation",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrganisationSummary"
              }
            }
          }
        },
        "responses": {
          "201": {
            "headers": {
              "API-Version": {
                "$ref": "#/components/headers/APIVersion"
              }
            },
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganisationSummary"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ProblemJson": {
        "title": "Problem JSON",
        "description": "Problem JSON schema representing errors and status code",
        "type": "object",
        "properties": {
          "status": {
            "type": "integer",
            "description": "The HTTP status code generated by the origin server for this occurrence of the problem",
            "example": 400
          },
          "title": {
            "type": "string",
            "description": "A short, human-readable summary of the problem type",
            "example": "Request validation failed"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "description": "List of errors encountered during processing",
              "properties": {
                "in": {
                  "type": "string",
                  "description": "Location of the error (e.g., body, query, header)",
                  "enum": [
                    "body",
                    "query"
                  ]
                },
                "location": {
                  "type": "string",
                  "description": "Location in the document where the error occurred (JSON Pointer)",
                  "example": "#/foo[0]/bar"
                },
                "code": {
                  "type": "string",
                  "description": "A code representing the type of error",
                  "example": "date.format"
                },
                "detail": {
                  "type": "string",
                  "description": "A detailed message describing the error",
                  "example": "must be ISO 8601"
                }
              },
              "required": [
                "in",
                "location",
                "code",
                "detail"
              ]
            }
          }
        },
        "required": [
          "status",
          "title"
        ]
      },
      "ResourceId": {
        "type": "string",
        "description": "Unique identifier of the schema",
        "examples": [
          "demo-bier"
        ]
      },
      "OrganisationUri": {
        "title": "Organisation URI",
        "description": "The unique identifier for an organisation",
        "type": "string",
        "format": "uri",
        "example": "https://developer.overheid.nl"
      },
      "SchemaUrl": {
        "type": "string",
        "format": "uri",
        "description": "Location of the JSON Schema document",
        "examples": [
          "https://schemas.example.org/Bier.schema.json"
        ]
      },
      "SchemaTitle": {
        "type": "string",
        "description": "Title of the JSON Schema (taken from the schema's title keyword)",
        "examples": [
          "Bier"
        ]
      },
      "SchemaDescription": {
        "type": "string",
        "description": "Description of the JSON Schema (taken from the schema's description keyword)",
        "examples": [
          "Het specifieke bier dat een brouwerij produceert"
        ]
      },
      "SchemaDialect": {
        "type": "string",
        "description": "Normalised JSON Schema dialect, derived from the $schema keyword",
        "enum": [
          "2020-12",
          "2019-09",
          "draft-07",
          "draft-06",
          "draft-04",
          "oas-3.1",
          "unknown"
        ],
        "examples": [
          "2020-12"
        ]
      },
      "SchemaRootType": {
        "type": "string",
        "description": "Root type of the JSON Schema, derived from the top-level type keyword",
        "enum": [
          "object",
          "array",
          "string",
          "number",
          "integer",
          "boolean",
          "null",
          "unknown"
        ],
        "examples": [
          "object"
        ]
      },
      "SchemaCollection": {
        "type": "string",
        "description": "Optional collection the schema belongs to",
        "examples": [
          "demo"
        ]
      },
      "CreatedAt": {
        "type": "string",
        "format": "date-time",
        "readOnly": true
      },
      "LastCrawledAt": {
        "type": "string",
        "format": "date-time",
        "readOnly": true
      },
      "OrganisationSummary": {
        "title": "Organisation summary",
        "description": "An organisation from the catalog",
        "type": "object",
        "properties": {
          "uri": {
            "$ref": "#/components/schemas/OrganisationUri"
          },
          "label": {
            "description": "The label of the organisation",
            "type": "string",
            "example": "developer.overheid.nl"
          }
        },
        "required": [
          "uri",
          "label"
        ]
      },
      "Contact": {
        "type": "object",
        "description": "Contact information of the schema maintainer",
        "properties": {
          "name": {
            "type": "string",
            "examples": [
              "Team developer.overheid.nl"
            ]
          },
          "email": {
            "type": "string",
            "format": "email",
            "examples": [
              "developer.overheid@geonovum.nl"
            ]
          },
          "url": {
            "type": "string",
            "format": "uri",
            "examples": [
              "https://developer.overheid.nl"
            ]
          }
        },
        "required": [
          "name"
        ]
      },
      "SchemaSummary": {
        "type": "object",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/ResourceId"
          },
          "schemaUrl": {
            "$ref": "#/components/schemas/SchemaUrl"
          },
          "title": {
            "$ref": "#/components/schemas/SchemaTitle"
          },
          "description": {
            "$ref": "#/components/schemas/SchemaDescription"
          },
          "dialect": {
            "$ref": "#/components/schemas/SchemaDialect"
          },
          "rootType": {
            "$ref": "#/components/schemas/SchemaRootType"
          },
          "collection": {
            "$ref": "#/components/schemas/SchemaCollection"
          },
          "contact": {
            "$ref": "#/components/schemas/Contact"
          },
          "organisation": {
            "$ref": "#/components/schemas/OrganisationSummary"
          },
          "createdAt": {
            "$ref": "#/components/schemas/CreatedAt"
          },
          "lastCrawledAt": {
            "$ref": "#/components/schemas/LastCrawledAt"
          }
        },
        "required": [
          "id",
          "title"
        ]
      },
      "SchemaDetail": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SchemaSummary"
          },
          {
            "type": "object",
            "properties": {
              "content": {
                "type": "object",
                "description": "The stored JSON Schema document"
              }
            }
          }
        ]
      },
      "SchemaPost": {
        "type": "object",
        "description": "Register a JSON Schema by URL or by providing the schema body directly. Either schemaUrl or schemaBody is required.",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/ResourceId"
          },
          "schemaUrl": {
            "$ref": "#/components/schemas/SchemaUrl"
          },
          "schemaBody": {
            "type": "string",
            "description": "The JSON Schema document itself, as a JSON string. Required when schemaUrl is omitted."
          },
          "organisationUri": {
            "$ref": "#/components/schemas/OrganisationUri"
          },
          "contact": {
            "$ref": "#/components/schemas/Contact"
          }
        },
        "required": [
          "organisationUri",
          "contact"
        ]
      },
      "FilterOption": {
        "type": "object",
        "description": "A single option within a multi-select filter group.",
        "required": [
          "value",
          "label",
          "count",
          "selected"
        ],
        "properties": {
          "value": {
            "type": "string",
            "description": "The filter value to pass as a query parameter."
          },
          "label": {
            "type": "string",
            "description": "Human-readable label for the option."
          },
          "description": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional description of the filter value."
          },
          "count": {
            "type": "integer",
            "description": "Number of repositories matching this option, given the other active filters."
          },
          "selected": {
            "type": "boolean",
            "description": "Whether this option is currently selected."
          }
        }
      },
      "FilterGroup": {
        "type": "object",
        "description": "A filter group with metadata and available options or current value.",
        "required": [
          "key",
          "label",
          "description",
          "type"
        ],
        "properties": {
          "key": {
            "type": "string",
            "description": "Unique key of the filter group, matching the query parameter name.",
            "example": "softwareType"
          },
          "label": {
            "type": "string",
            "description": "Human-readable label for the filter group."
          },
          "description": {
            "type": "string",
            "description": "Description of what this filter does."
          },
          "type": {
            "type": "string",
            "enum": [
              "toggle",
              "date",
              "multi-select",
              "single-select"
            ],
            "description": "Filter type. toggle: on/off switch. date: date input (ISO 8601). multi-select: multiple values selectable. single-select: one value selectable from a list."
          },
          "value": {
            "oneOf": [
              {
                "type": "boolean"
              },
              {
                "type": "string"
              }
            ],
            "description": "Current filter value. For toggle: boolean true or false. For date: ISO 8601 date (yyyy-MM-dd). Not present for multi-select."
          },
          "count": {
            "type": [
              "integer",
              "null"
            ],
            "description": "For toggle filters, the number of repositories with the toggle condition available. For date filters, the number of repositories matching the current date value; only present when a date is set. Not present for multi-select."
          },
          "options": {
            "type": "array",
            "description": "Available options with counts. Only present for multi-select type.",
            "items": {
              "$ref": "#/components/schemas/FilterOption"
            }
          }
        }
      }
    },
    "parameters": {
      "Page": {
        "name": "page",
        "in": "query",
        "required": false,
        "description": "Page number (1-based).",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "default": 1
        }
      },
      "PerPage": {
        "name": "perPage",
        "in": "query",
        "required": false,
        "description": "Number of results per page.",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 100,
          "default": 20
        }
      },
      "SearchFilter": {
        "name": "q",
        "in": "query",
        "required": false,
        "description": "Search term to combine with repository filters. Matches repository name, short description, long description, publiccode.yml url and publiccode.yml landingURL.",
        "schema": {
          "type": "string"
        }
      },
      "OrganisationFilter": {
        "name": "organisation",
        "in": "query",
        "required": false,
        "description": "Filter by organisation URI.",
        "schema": {
          "type": "string",
          "format": "uri"
        }
      },
      "DialectFilter": {
        "name": "dialect",
        "in": "query",
        "description": "Filter schemas by JSON Schema dialect. Repeat the parameter to filter on multiple dialects.",
        "schema": {
          "type": "array",
          "items": {
            "$ref": "#/components/schemas/SchemaDialect"
          }
        },
        "explode": true
      },
      "RootTypeFilter": {
        "name": "rootType",
        "in": "query",
        "description": "Filter schemas by root type. Repeat the parameter to filter on multiple root types.",
        "schema": {
          "type": "array",
          "items": {
            "$ref": "#/components/schemas/SchemaRootType"
          }
        },
        "explode": true
      },
      "ResourceId": {
        "name": "id",
        "in": "path",
        "required": true,
        "description": "Unique identifier of the schema",
        "schema": {
          "$ref": "#/components/schemas/ResourceId"
        }
      }
    },
    "headers": {
      "APIVersion": {
        "description": "Semver of this API",
        "schema": {
          "type": "string",
          "example": "1.0.0",
          "pattern": "^(0|[1-9]\\d*)\\.(0|[1-9]\\d*)\\.(0|[1-9]\\d*)(?:-((?:0|[1-9]\\d*|\\d*[a-zA-Z-][0-9a-zA-Z-]*)(?:\\.(?:0|[1-9]\\d*|\\d*[a-zA-Z-][0-9a-zA-Z-]*))*))?(?:\\+([0-9a-zA-Z-]+(?:\\.[0-9a-zA-Z-]+)*))?$",
          "externalDocs": {
            "description": " /core/version-header: Return the full version number in a response header",
            "url": "https://logius-standaarden.github.io/API-Design-Rules/#/core/version-header"
          }
        }
      },
      "Link": {
        "description": "Links to the previous, next, last or first pages",
        "schema": {
          "type": "string",
          "example": "<https://api.developer.overheid.nl/api-register/v1/apis?page=2>; rel='prev', <https://api.developer.overheid.nl/api-register/v1/apis?page=3>; rel='next'",
          "externalDocs": {
            "description": "W3C reference",
            "url": "https://www.w3.org/wiki/LinkHeader"
          }
        }
      },
      "TotalCount": {
        "description": "Total number of items available",
        "schema": {
          "type": "integer",
          "example": 124
        }
      },
      "CurrentPage": {
        "description": "Current page number",
        "schema": {
          "type": "integer",
          "example": 1
        }
      },
      "PerPage": {
        "description": "Number of items per page",
        "schema": {
          "type": "integer",
          "example": 10
        }
      },
      "TotalPages": {
        "description": "Total number of pages available",
        "schema": {
          "type": "integer",
          "example": 13
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Bad request",
        "headers": {
          "API-Version": {
            "$ref": "#/components/headers/APIVersion"
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemJson"
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource does not exist",
        "headers": {
          "API-Version": {
            "$ref": "#/components/headers/APIVersion"
          }
        },
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/ProblemJson"
            }
          }
        }
      }
    },
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "name": "X-Api-Key",
        "in": "header",
        "description": "Read-only API keys for public endpoints can be requested at https://apis.developer.overheid.nl/apis/key-aanvragen."
      },
      "clientCredentials": {
        "type": "oauth2",
        "flows": {
          "clientCredentials": {
            "scopes": {
              "schemas:read": "Read access to schemas",
              "schemas:write": "Write access to schemas",
              "organisations:read": "Read access to organisations",
              "organisations:write": "Write access to organisations"
            },
            "tokenUrl": "https://auth.developer.overheid.nl/realms/don/protocol/openid-connect/token"
          }
        }
      }
    }
  }
}